+++ src/components/VictoryPage.tsx
@@ -1,12 +1,7 @@
 import React from "react";
 import "./VictoryPage.css";
 import spotifyLogo from "../assets/Spotify_Logo_RGB_Black.png";
-
-type MatchSet = Array<{
-  matchRound: number;
-  matchComponent: React.JSX.Element;
-  matchWinnerId: string | null;
-}>;
+import { MatchSet } from "../types/index";
 
 function roundToDivs(matchesInRound: MatchSet): React.JSX.Element[] {
   const divArray = [];
