+++ src/types/MatchSet.type.ts
@@ -1,0 +1,7 @@
+type MatchSet = Array<{
+    matchRound: number;
+    matchComponent: React.JSX.Element;
+    matchWinnerId: string | null;
+}>;
+
+export {MatchSet};
