+++ src/types/Song.type.ts
@@ -1,0 +1,7 @@
+type Song = {
+    id: string;
+    artists: Array<{ name: string }>;
+    name: string;
+};
+
+export {Song};
