+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./Song.type"
+export * from "./MatchSet.type"
